--- .husky/pre-push
@@ -1,2 +1,0 @@
-#!/bin/sh
-npm test
