+++ src/components/pages/DetailedPetPage/DetailedPetPage.module.css
