+++ src/components/pages/MainPage/SelectedItemsManager/SelectedItemsManager.module.css
