+++ src/components/pages/MainPage/TopControls/TopControls.types.ts
